requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pymongo==4.3.3
requests==2.31.0
python-telegram-bot==20.5
gunicorn==20.1.0
